ai_service/requirements.txt
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
matplotlib==3.7.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.7 